Move view mixins to their own source file
